Test inserts, commits and rollbacks. Test inserting text and binary blobs. Fixed bug inserting blobs larger than 64K.
git-svn-id: https://svn.ttiltd.com/svn/dev.rb/trunk/libraries/fb@83 e6efdd46-4312-0410-9032-dd8a18d5a4db

+++ fb.c
@@ -930,7 +930,7 @@ static void fb_cursor_set_inputparams(struct FbCursor *fb_cursor, int argc, VALU
 	static char blob_items[] = { isc_info_blob_max_segment };
 	char blob_info[16];
 	char *p;
-	unsigned short length;
+	long length;
 	struct time_object *tobj;
 
 	Data_Get_Struct(fb_cursor->connection, struct FbConnection, fb_connection);
